pyproject.toml
[build-system]
requires = ["hatchling>=1.27.0"]
build-backend = "hatchling.build"

[project]
name = "axiomdoc"
version = "1.0.1"
description = "Open-source document intelligence for extraction, structure preservation, XML export, and RAG indexing."
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
authors = [
  { name = "Meet Jethwa" }
]
maintainers = [
  { name = "Meet Jethwa" }
]
keywords = [
  "document-intelligence",
  "rag",
  "pdf",
  "xml",
  "docx",
  "xlsx",
  "information-extraction"
]
classifiers = [
  "Development Status :: 5 - Production/Stable",
  "Intended Audience :: Developers",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Topic :: Scientific/Engineering :: Artificial Intelligence",
  "Topic :: Software Development :: Libraries :: Python Modules"
]
dependencies = []

[project.urls]
Homepage = "https://github.com/Meet2147/axiomdoc"
Repository = "https://github.com/Meet2147/axiomdoc.git"
Issues = "https://github.com/Meet2147/axiomdoc/issues"

[project.optional-dependencies]
pdf = [
  "pymupdf>=1.25.0",
  "pillow>=10.4.0",
]
docx = [
  "python-docx>=1.1.2",
]
xlsx = [
  "openpyxl>=3.1.5",
]
xml = [
  "lxml>=5.3.0",
]
full = [
  "pymupdf>=1.25.0",
  "pillow>=10.4.0",
  "python-docx>=1.1.2",
  "openpyxl>=3.1.5",
  "lxml>=5.3.0",
]
dev = [
  "pytest>=8.3.0",
]

[project.scripts]
axiomdoc = "axiomdoc.cli:main"

[tool.hatch.build]
include = [
  "src/axiomdoc/**/*.py",
  "README.md",
  "LICENSE",
  "docs/**",
  "assets/**",
]

[tool.hatch.build.targets.wheel]
packages = ["src/axiomdoc"]

[tool.hatch.build.targets.sdist]
include = [
  "/src",
  "/README.md",
  "/LICENSE",
  "/assets",
  "/docs",
  "/tests",
  "/pyproject.toml"
]

[tool.pytest.ini_options]
testpaths = ["tests"]
